requirements.txt
# AWS SDK for Python
boto3>=1.34.0

# Environment
python-dotenv>=1.0.0

# HTTP client (used by a2a client)
httpx>=0.27.0

# Strands agents
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
strands-agents-builder

# Strands agents with A2A support
strands-agents[a2a]